requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.107.3
openai-whisper>=20231117
gtts==2.5.1
numpy==2.2.6
soundfile==0.13.1
python-dotenv==1.1.1
pydantic==2.5.0
python-multipart==0.0.6
psycopg2-binary==2.9.7